requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
idna==2.10
oauthlib==3.1.0
PySocks==1.7.1
python-dotenv==0.15.0
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.2
tweepy==3.10.0
urllib3==1.26.3